split tests in two files
